ADD first iteraton of cute dashbee.
pretty happy with the result. missing fiew components from the 'trending', but hopefully i already got the gist of the elements and items, just missing the time to actually add them

pretty happy to be able to add more, eg. tooltips, colouring the svgs inside CSS, etc.
pretty happy to use variables for colours, not very impactful but hopefully a good habbit

still having difficulties with grid flexbox, in some point,
still having a LOT of difficulties with making a FIXED sidebar.
still confused on using UNITS rem em etc, have to revisit some of the ideas.

unlikely to revisit though, as time is really not in my sides